Non-hardcoded boilerplate directory; renaming directory arguments.
Boilerplate directory is now specified in the config file. This was
hardcoded pretty deep so a few functions dealing with the boilerplate
directory have been removed. The net result of this is more calls to
os.path.join, but I think they're all reasonable in context.

Finally, functions that previously took a boilerplate directory as an
argument now take both an output_dir  and the name of the boilerplate
directory. This leads to slightly more repetition in the code, but
makes the interfaces of the functions considerably more consistent.

R=mgiuca@google.com

CL=329417013