Stubbed runtime and app.runtime in caterpillar.js. Resolves #8.
Currently, caterpillar.py stubs app.runtime so that the initial
call to onLaunched for window creation won't fail. If that call
fails, then the service worker won't install, as it has a
script error. This CL moves this stubbing to caterpillar.js,
which is always a dependency. This is an improvement over the
existing location of the stub, since the stub is common code
and static, and so should not be included in a generated file.

Also, currently the runtime polyfill is required as a
dependency for caterpillar.js, since caterpillar.setError needs
to access the runtime namespace. This CL adds a runtime stub to
caterpillar.js so runtime is no longer needed as a dependency.
This resolves issue #8.

https://github.com/chromium/caterpillar/issues/8

R=raymes@chromium.org

Review URL: https://codereview.chromium.org/1642833002 .