Fixes Caterpillar crashing while generating report.
Caterpillar currently crashes while generating a report. This
is due to API changes between writing and merging the report
linking branch. This CL updates the report code to account for
the new API and new variable names.

Tests to catch the problem if it arises again in future are
also added in this CL.

Resolves #42. https://github.com/chromium/caterpillar/issues/42

R=mgiuca@chromium.org

Review URL: https://codereview.chromium.org/1662833005 .