report.format_html now ignores APIs not passed to it.
Previously, report.format_html would error if it tried to highlight
something that looked like an API, but didn't show up in the APIs dict
passed to it as an argument. For example, "chrome.google.com" would
trigger this bug. This CL changes the behaviour such that if something
API-like isn't contained in the APIs dict passed as an argument, then
report.format_html will assume that it is not an API, and will ignore
it.

Resolves https://github.com/chromium/caterpillar/issues/43 .

R=mgiuca@chromium.org

Review URL: https://codereview.chromium.org/1677943002 .