Errors raised in setup_output_dir are no longer OSErrors.
This CL changes the way that errors are logged for setup_output_dir
such that a custom error type is logged instead of an OSError.
This means that we are not showing OSErrors to end users, and we no
longer show blank errors for input directory errors.

Resolves https://github.com/chromium/caterpillar/issues/35 .

R=mgiuca@chromium.org

Review URL: https://codereview.chromium.org/1679733002 .