Ignore GET parameters in service worker cache lookups.
GET parameters were previously significant, so while the service worker
may have cached assets/resource.css, the Chrome App may request
assets/resource.css?v=1212 and miss the resource. This CL ignores GET
parameters when performing cache lookups so that this does not occur,
since the cache is manually managed by the service worker.

To ignore the GET parameters, I've written a helper function
matchIgnoreParams. This is not necessary by the spec, which includes an
option ignoreSearch in the parameters of Cache.match for exactly this
reason, but ignoreSearch is currently not supported by Chrome. Future
versions of the service worker should use ignoreSearch instead.

Included also in this CL are fixes to code style in sw_static.js. This
brings sw_static.js into line with the rest of the JavaScript in the
project.

Resolves https://github.com/chromium/caterpillar/issues/29 .

R=raymes@chromium.org

Review URL: https://codereview.chromium.org/1684503002 .