Replaced end-to-end test setup method with a setup classmethod.
All of the end-to-end tests are read only, so reconverting the test
Chrome App every test is not necessary. Converting is also slow due to
the installation of dependencies (through bower) and as far as I can
tell this can't be stubbed out. This CL changes the end-to-end tests
such that Caterpillar is only run once. The net effect of this is that
the end-to-end test is ~3x faster.

Other options for speeding this up would include somehow stubbing bower
(I don't think this is possible) or combining all of the read-only tests
into one test function (which I think is messy given they do different
things).

R=mgiuca@chromium.org

Review URL: https://codereview.chromium.org/1683553004 .